1st commit - JavierNieto_PersistentDataApp
You will be creating an app that saves a few items of data that you select on your phone. The choice of the data will be yours, but it should be something that makes sense to store permanently on your phone. You may also select the method that makes the most sense to you.